+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 # Changelog
 
 * Feat: Add more convenient way to pass BeforeSpanCallback in OpenFeign integration (#1637)
+* Bump: sentry-native to 0.4.12 (#1651)
 
 ## Unreleased